# src/test/resources/api/openapi.yaml
openapi: 3.0.1
info:
  title: api-oep-integrator
  contact: {}
  license:
    name: MIT License
    url: https://opensource.org/licenses/MIT
  version: "1.0"
servers:
- url: http://localhost:8080
  description: Generated server url
...
paths:
...
  /{municipalityId}/{instanceType}/webmessages:
    post:
      tags:
      - Webmessages
      summary: Create webmessage
      operationId: createWebmessage
      parameters:
      - name: municipalityId
        in: path
        description: Municipality id
        required: true
        schema:
          type: string
        example: 2281
      - name: instanceType
        in: path
        description: Which instanceType a message should be sent to
        required: true
        schema:
          $ref: "#/components/schemas/InstanceType"
        example: INTERNAL
      requestBody:
        content:
          multipart/form-data:
            schema:
              required:
              - request
              type: object
              properties:
                request:
                  $ref: "#/components/schemas/WebmessageRequest"
                attachments:
                  type: array
                  items:
                    type: string
                    format: binary
      responses:
        "201":
          description: Successful operation
          headers:
            Location:
              style: simple
              schema:
                type: string
        "500":
          description: Internal Server error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "502":
          description: Bad gateway
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "400":
          description: Bad request
          content:
            application/problem+json:
              schema:
                oneOf:
                - $ref: "#/components/schemas/Problem"
                - $ref: "#/components/schemas/ConstraintViolationProblem"
...
  /{municipalityId}/{instanceType}/webmessages/families/{familyId}:
    get:
      tags:
      - Webmessages
      summary: Get all webmessages for a familyId
      operationId: getWebmessagesByFamilyId
      parameters:
      - name: municipalityId
        in: path
        description: Municipality id
        required: true
        schema:
          type: string
        example: 2281
      - name: instanceType
        in: path
        description: Which instanceType a message should be sent to
        required: true
        schema:
          $ref: "#/components/schemas/InstanceType"
        example: INTERNAL
      - name: familyId
        in: path
        description: Family id
        required: true
        schema:
          type: string
        example: 123
      - name: fromDateTime
        in: query
        description: The start date and time for filtering web messages (optional)
        required: false
        schema:
          type: string
          format: date-time
        example: 2024-01-31T12:00:00
      - name: toDateTime
        in: query
        description: The end date and time for filtering web messages (optional).
        required: false
        schema:
          type: string
          format: date-time
        example: 2024-01-31T12:00:00
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Webmessage"
        "500":
          description: Internal Server error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "502":
          description: Bad gateway
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "400":
          description: Bad request
          content:
            application/problem+json:
              schema:
                oneOf:
                - $ref: "#/components/schemas/Problem"
                - $ref: "#/components/schemas/ConstraintViolationProblem"
  /{municipalityId}/{instanceType}/webmessages/cases/{flowInstanceId}:
    get:
      tags:
      - Webmessages
      summary: Get all webmessages for a given case (errand)
      operationId: getWebmessagesByFlowInstanceId
      parameters:
      - name: municipalityId
        in: path
        description: Municipality id
        required: true
        schema:
          type: string
        example: 2281
      - name: instanceType
        in: path
        description: Which instanceType a message should be sent to
        required: true
        schema:
          $ref: "#/components/schemas/InstanceType"
        example: INTERNAL
      - name: flowInstanceId
        in: path
        description: Flow instance id
        required: true
        schema:
          type: string
        example: 123
      - name: fromDateTime
        in: query
        description: The start date and time for filtering web messages (optional)
        required: false
        schema:
          type: string
          format: date-time
        example: 2024-01-31T12:00:00
      - name: toDateTime
        in: query
        description: The end date and time for filtering web messages (optional).
        required: false
        schema:
          type: string
          format: date-time
        example: 2024-01-31T12:00:00
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Webmessage"
        "500":
          description: Internal Server error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "502":
          description: Bad gateway
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "400":
          description: Bad request
          content:
            application/problem+json:
              schema:
                oneOf:
                - $ref: "#/components/schemas/Problem"
                - $ref: "#/components/schemas/ConstraintViolationProblem"
  /{municipalityId}/{instanceType}/webmessages/attachments/{attachmentId}:
    get:
      tags:
      - Webmessages
      summary: Get attachment by id
      operationId: getAttachmentById
      parameters:
      - name: municipalityId
        in: path
        description: Municipality id
        required: true
        schema:
          type: string
        example: 2281
      - name: instanceType
        in: path
        description: Which instanceType a message should be sent to
        required: true
        schema:
          $ref: "#/components/schemas/InstanceType"
        example: INTERNAL
      - name: attachmentId
        in: path
        description: Attachment id
        required: true
        schema:
          type: integer
          format: int32
        example: 123
      responses:
        "200":
          description: Successful operation
        "500":
          description: Internal Server error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "502":
          description: Bad gateway
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "400":
          description: Bad request
          content:
            application/problem+json:
              schema:
                oneOf:
                - $ref: "#/components/schemas/Problem"
                - $ref: "#/components/schemas/ConstraintViolationProblem"
...
components:
  schemas:
    Problem:
      type: object
      properties:
        instance:
          type: string
          format: uri
        type:
          type: string
          format: uri
        parameters:
          type: object
          additionalProperties:
            type: object
        status:
          $ref: "#/components/schemas/StatusType"
        detail:
          type: string
        title:
          type: string
    StatusType:
      type: object
      properties:
        reasonPhrase:
          type: string
        statusCode:
          type: integer
          format: int32
    InstanceType:
      type: string
      description: InstanceType type. INTERNAL/EXTERNAL
      example: INTERNAL
      enum:
      - INTERNAL
      - EXTERNAL
...
    ConstraintViolationProblem:
      type: object
      properties:
        cause:
          $ref: "#/components/schemas/ThrowableProblem"
        stackTrace:
          type: array
          items:
            type: object
            properties:
              classLoaderName:
                type: string
              moduleName:
                type: string
              moduleVersion:
                type: string
              methodName:
                type: string
              fileName:
                type: string
              lineNumber:
                type: integer
                format: int32
              className:
                type: string
              nativeMethod:
                type: boolean
        type:
          type: string
          format: uri
        status:
          $ref: "#/components/schemas/StatusType"
        violations:
          type: array
          items:
            $ref: "#/components/schemas/Violation"
        title:
          type: string
        message:
          type: string
        instance:
          type: string
          format: uri
        parameters:
          type: object
          additionalProperties:
            type: object
        detail:
          type: string
        suppressed:
          type: array
          items:
            type: object
            properties:
              stackTrace:
                type: array
                items:
                  type: object
                  properties:
                    classLoaderName:
                      type: string
                    moduleName:
                      type: string
                    moduleVersion:
                      type: string
                    methodName:
                      type: string
                    fileName:
                      type: string
                    lineNumber:
                      type: integer
                      format: int32
                    className:
                      type: string
                    nativeMethod:
                      type: boolean
              message:
                type: string
              localizedMessage:
                type: string
        localizedMessage:
          type: string
    ThrowableProblem:
      type: object
      properties:
        cause:
          $ref: "#/components/schemas/ThrowableProblem"
        stackTrace:
          type: array
          items:
            type: object
            properties:
              classLoaderName:
                type: string
              moduleName:
                type: string
              moduleVersion:
                type: string
              methodName:
                type: string
              fileName:
                type: string
              lineNumber:
                type: integer
                format: int32
              className:
                type: string
              nativeMethod:
                type: boolean
        message:
          type: string
        instance:
          type: string
          format: uri
        type:
          type: string
          format: uri
        parameters:
          type: object
          additionalProperties:
            type: object
        status:
          $ref: "#/components/schemas/StatusType"
        detail:
          type: string
        title:
          type: string
        suppressed:
          type: array
          items:
            type: object
            properties:
              stackTrace:
                type: array
                items:
                  type: object
                  properties:
                    classLoaderName:
                      type: string
                    moduleName:
                      type: string
                    moduleVersion:
                      type: string
                    methodName:
                      type: string
                    fileName:
                      type: string
                    lineNumber:
                      type: integer
                      format: int32
                    className:
                      type: string
                    nativeMethod:
                      type: boolean
              message:
                type: string
              localizedMessage:
                type: string
        localizedMessage:
          type: string
    Violation:
      type: object
      properties:
        field:
          type: string
        message:
          type: string
...
    ExternalReference:
      type: object
      properties:
        key:
          type: string
          description: The external reference key
          example: flowInstanceId
        value:
          type: string
          description: The external reference value
          example: "356746349"
      description: ExternalReference model
    Sender:
      type: object
      properties:
        userId:
          type: string
          description: The user ID of the sender. I.e. employee ID
          example: joe01doe
        administratorId:
          type: string
          description: The user ID of the case administrator. I.e. the employee ID
            of the administrative official.
          example: joe01doe
        partyId:
          type: string
          description: The party ID of the sender.
          example: 582462d3-ae94-476f-96de-f495a927dea8
      description: "Sender model. The attributes in this object are mutually exclusive,\
        \ meaning that only one of them can be set at a time."
    WebmessageRequest:
      required:
      - message
      - sender
      type: object
      properties:
        sender:
          $ref: "#/components/schemas/Sender"
        message:
          type: string
          description: The message
          example: This is a message
        externalReferences:
          type: array
          items:
            $ref: "#/components/schemas/ExternalReference"
      description: Webmessage request model
...
    Direction:
      type: string
      description: Direction model
      example: INBOUND
      enum:
      - INBOUND
      - OUTBOUND
    Webmessage:
      type: object
      properties:
        id:
          type: integer
          description: The webMessageCollector Id for the message
          format: int32
          example: 1
        direction:
          $ref: "#/components/schemas/Direction"
        municipalityId:
          type: string
          description: The municipality id
          example: "2281"
        familyId:
          type: string
          description: What E-service the message was found in
          example: "501"
        externalCaseId:
          type: string
          description: 'The external caseID '
          example: caa230c6-abb4-4592-ad9a-34e263c2787b
        message:
          type: string
          description: 'The message '
          example: Hello World
        messageId:
          type: string
          description: The unique messageId from openE for the message
          example: "12"
        sent:
          type: string
          description: 'Time and date the message was sent '
          format: date-time
        username:
          type: string
          description: Username for the poster
          example: te01st
        firstName:
          type: string
          description: 'Firstname of the poster '
          example: Test
        lastName:
          type: string
          description: Lastname of the poster
          example: Testsson
        email:
          type: string
          description: Email for the poster
          example: test@sundsvall.se
        userId:
          type: string
          description: The userId for the poster
          example: "123"
        attachments:
          type: array
          items:
            $ref: "#/components/schemas/WebmessageAttachment"
        instance:
          type: string
          description: The instance of the message
          example: external
      description: Webmessage model
    WebmessageAttachment:
      type: object
      properties:
        attachmentId:
          type: integer
          description: The Id for the attachment
          format: int32
          example: 1
        name:
          type: string
          description: The name of the file
          example: file.txt
        extension:
          type: string
          description: The extension of the file
          example: txt
        mimeType:
          type: string
          description: The mime type of the file
          example: text/plain
      description: Webmessage attachment model
      example: "attachment1, attachment2"
...
  securitySchemes: {}